Turn upload into assets module
SDESK-7286
