+++ newsroom/news_api/default_settings.py
@@ -33,7 +33,6 @@ CORE_APPS = [
     "newsroom.products",
     "newsroom.news_api.api_audit",
     "newsroom.news_api.news.assets.assets",
-    "newsroom.upload",
     "newsroom.news_api.news.atom.atom",
     "newsroom.history",
 ]
