+++ newsroom/web/default_settings.py
@@ -176,7 +176,7 @@ CORE_APPS = [
     "newsroom.notifications.send_scheduled_notifications",
 ]
 
-MODULES = ["newsroom.ui_config_async", "newsroom.upload"]
+MODULES = ["newsroom.ui_config_async", "newsroom.assets.module"]
 
 SITE_NAME = "Newshub"
 COPYRIGHT_HOLDER = "Sourcefabric"
