+++ newsroom/wire/views.py
@@ -1,4 +1,5 @@
 import io
+import asyncio
 import zipfile
 import superdesk
 
@@ -56,7 +57,8 @@ from newsroom.public.views import (
 
 from .search import get_bookmarks_count
 from .items import get_items_for_dashboard
-from ..upload import ASSETS_RESOURCE, get_upload
+from newsroom.assets.module import ASSETS_RESOURCE
+from newsroom.assets.views import get_upload
 from newsroom.ui_config_async import UiConfigResourceService
 from newsroom.users import get_user_profile_data
 
@@ -311,7 +313,8 @@ def download():
         if len(items) == 1:
             try:
                 picture = formatter.format_item(items[0], item_type=item_type)
-                return get_upload(picture["media"], filename="baseimage%s" % picture["file_extension"])
+                # TODO: running through `asyncio.run` until we migrate this view
+                return asyncio.run(get_upload(picture["media"], filename="baseimage%s" % picture["file_extension"]))
             except ValueError:
                 return abort(404)
         else:
