+++ tests/assets/test_views.py
